// pos-Frontend/src/app/model/StockDetail.ts
export class StockDetail {
  intStockDetailId?: number;
  intStid?: number;
  varItemName?: string;
  intItemId?: number;
  intWarehouseId?: number;
  intQuantity?: number;
  dcRate?: number;
  dcAmount?: number;
  dcDisc?: number;
  dcDiscAmount?: number;
  dcExclTaxAmount?: number;
  dcTax?: number;
  dcTaxAmount?: number;
  dcInclTaxAmount?: number;
  varType?: string;
  dcPurRate?: number;
  dtCreationDate?: Date;
  dtUpdationDate?: Date;
  intCreatedBy?: number;
  intUpdatedBy?: number;
}